[Chore] Module Extension 분리, Module Sendable 추가
