+++ juinjang/Tuist/ProjectDescriptionHelpers/Target+Extensions.swift
@@ -22,11 +22,11 @@ extension Target {
             name: name ?? module.moduleSubName,
             destinations: Environment.destinations,
             product: product,
-            productName: productName ??  "\(module.moduleName)\(module.moduleSubName)",
-            bundleId: bundleId ?? "com.juinjangteam.juinjang.\(module.moduleName.lowercased()).\(module.moduleSubName)",
+            productName: productName ?? module.defaultProductName,
+            bundleId: bundleId ?? module.defaultBundleID,
             deploymentTargets: Environment.deploymentTarget,
             infoPlist: infoPlist,
-            sources: sources ?? ["\(module.moduleSubName)/**"],
+            sources: sources ?? module.defaultSources,
             resources: resources,
             entitlements: entitlements,
             dependencies: dependencies.map {
